+++ 61+蒋明睿/README.md
@@ -1,1 +1,2 @@
 
+第二周作业，四维向量根据排序三分类，目前面临的问题是训练结果比较随机，时好时坏
